More movement changes
Mostly buried for now but hoping to bust them out when we have a
backwards walking animation. Very experimental, but trying to prevent
the player from accelerating in any direction unless they have a foot on
the ground, and only able to accelerate backwards if they have a
grounded foot in front of them and vice versa. This is the first step in
what I am hoping will reduce ridiculous sliding, like drag cars.