Re-jiggered metabolism
Digestion rate is now an average of the contents of the stomach.
Contents move from stomach to gut. I am working on another class that
will wait for certain periods of time before moving the contents of the
player's gut to the bloodstream. This will allow for different foods
getting into the bloodstream at different rates and times.