More metabolism logic
Wow, what a wrestle. Food moves from stomach to gut at a rate derived
from the proportions of each food type in the stomach via a list, where
it is held for a set amount of time and the current digestion rate for
that food is recorded. When the time comes for that food to be released,
the release rate of the food into the bloodstream is averaged between
what it is now and the rate of the new amount being released from the
gut, proportionally. This means that if you eat a variety of foods, it
will not be released from your gut into your blood stream faster than it
moves from your stomach into your gut.