Fixed using same list -> collection modified error
So most things in C# are references, and when you say List something =
somethingElse, they are now both references to the same list. I have
changed it to create a new List, which takes in the old list. This
particular way of doing it will copy references to every item in the old
list into the new one, which is probably the behaviour you wanted.
#programminglessons