Created system for leaning into motion direction
Added an empty scipt called AcclerationLeadObject.cs. Attempted to make
it work, but failed. The script is attached to LeanTarget, a child of
the player object. It is set up and ready to go to make the player lean
towards it, all that needs doing is the code to make the lean target
move correctly when the player moves.