Moved player animation to Blend Tree system
Changed character_test_anim_behavious.cs to PlayerAnimation.cs. Moved
player animations to the Blend tree system. Movement animations are now
controlled outside of scripts, and are dependant on the
PlayerAnimation.cs script setting the velForward and velStrafe
parameters in the animator object. This allows unity to take control of
when to move between different movement animations depending on player
speed, and massively simplifies the animation workflow. Only downside is
that transitions seem to be a bit broken now. Will look into that.