Inventory work + inventory is in front of all 3d objects
You can see items that have been picked up better now with the middle
click inventory. They are laid out in a slight grid pattern. At the
moment it looks weird sometimes because it restores things to their
original 0 rotation. Is it possible to have the models centred on the
actual centre of the models, and rotated to face up, and then rotate
them and position them in game?