Added item models: Apple, berry
Just some models. They should be appropriate scale. I tested it and one
square unit in blender = one square unit in Unity. If we go with that
being a meter, we should be able to make models that are scaled
correctly just by using their real world lengths etc when modelling them
in blender.